requirements.txt
opencv-python
scikit-learn
tensorflow
pyyaml
numpy